Updated search bar to be a component
