+++ src/styles/navbar.module.css
@@ -24,6 +24,7 @@
 
 .logo img {
   height: 30px;
+  margin: 0;
 }
 
 .links a {
