+++ src/styles/searchBar.module.css
@@ -1,0 +1,32 @@
+/* Search form styling */
+.searchForm {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 1rem;
+  margin-bottom: 2rem;
+}
+
+.searchForm input[type="text"] {
+  padding: 0.6rem 1rem;
+  border: 1px solid #ccc;
+  border-radius: 8px;
+  font-size: 1rem;
+  width: 100%;
+  max-width: 400px;
+}
+
+.searchForm button[type="submit"] {
+  padding: 0.6rem 1.2rem;
+  background-color: #1e1e2f;
+  color: white;
+  border: none;
+  border-radius: 8px;
+  font-weight: bold;
+  cursor: pointer;
+  transition: background-color 0.3s;
+}
+
+.searchForm button[type="submit"]:hover {
+  background-color: #2e2e4f;
+}
